+++ R/g.part2.R
@@ -18,9 +18,16 @@ g.part2 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
   params_output = params$params_output
   params_general = params$params_general
   #-----------------------------
+  use_qwindow_as_diary = TRUE # If there is a diary specified via qwindow use it as qwindow
   if (is.numeric(params_247[["qwindow"]])) {
     params_247[["qwindow"]] = params_247[["qwindow"]][order(params_247[["qwindow"]])]
   } else if (is.character(params_247[["qwindow"]])) {
+    if (length(grep(pattern = "onlyfilter|filteronly", x = params_247[["qwindow"]])) > 0) {
+      # Do not use diary specified for qwindow if it has the word
+      # "onlyfilter" or "filteronly", but use it for filterning nighttime nonwear
+      # note that this filtering is only use if parameter nonwearFiltermaxHours is specified.
+      use_qwindow_as_diary = FALSE 
+    }
     params_247[["qwindow"]] = g.conv.actlog(params_247[["qwindow"]],
                                             params_247[["qwindow_dateformat"]],
                                             epochSize = params_general[["windowsizes"]][1])
@@ -83,7 +90,7 @@ g.part2 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
                         myfun=c(), params_cleaning = c(), params_247 = c(),
                         params_phyact = c(), params_output = c(), params_general = c(),
                         path, ms2.out, foldername, fullfilenames, folderstructure, referencefnames,
-                        daySUMMARY, pdffilenumb, pdfpagecount, csvfolder, cnt78, verbose) {
+                        daySUMMARY, pdffilenumb, pdfpagecount, csvfolder, cnt78, verbose, use_qwindow_as_diary) {
     Nappended = I_list = tail_expansion_log =  NULL
     if (length(ffdone) > 0) {
       if (length(which(ffdone == as.character(unlist(strsplit(fnames[i], "eta_"))[2]))) > 0) {
@@ -150,13 +157,27 @@ g.part2 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
             }
           }
         }
+        qwindowImp = params_247[["qwindow"]]
+        if (use_qwindow_as_diary == FALSE) {
+          # reset qwindow to default, because it is only used
+          # for filtering short nighttime nonwear 
+          params_247[["qwindow"]] = c(0, 24)
+        }
+        if (inherits(qwindowImp, "data.frame")) {
+          qwindowImp = qwindowImp[which(qwindowImp$ID == ID),]
+          if (nrow(qwindowImp) == 0) {
+            qwindowImp = c(0, 6) # If participant not present in diary
+          }
+        } else {
+          qwindowImp = NULL
+        }
         IMP = g.impute(M, I,
                        params_cleaning = params_cleaning,
                        dayborder = params_general[["dayborder"]],
                        desiredtz = params_general[["desiredtz"]],
                        TimeSegments2Zero = TimeSegments2Zero,
                        acc.metric = params_general[["acc.metric"]],
-                       ID = ID)
+                       ID = ID, qwindowImp = qwindowImp)
         
         if (params_cleaning[["do.imp"]] == FALSE) { #for those interested in sensisitivity analysis
           IMP$metashort = M$metashort
@@ -315,7 +336,8 @@ g.part2 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
                                                   params_phyact, params_output, params_general,
                                                   path, ms2.out, foldername, fullfilenames,
                                                   folderstructure, referencefnames,
-                                                  daySUMMARY, pdffilenumb, pdfpagecount, csvfolder, cnt78, verbose)
+                                                  daySUMMARY, pdffilenumb, pdfpagecount, 
+                                                  csvfolder, cnt78, verbose, use_qwindow_as_diary)
                                        
                                      })
                                      return(tryCatchResult)
@@ -335,7 +357,8 @@ g.part2 = function(datadir = c(), metadatadir = c(), f0 = c(), f1 = c(),
                  params_phyact, params_output, params_general,
                  path, ms2.out, foldername, fullfilenames,
                  folderstructure, referencefnames,
-                 daySUMMARY, pdffilenumb, pdfpagecount, csvfolder, cnt78, verbose)
+                 daySUMMARY, pdffilenumb, pdfpagecount,
+                 csvfolder, cnt78, verbose, use_qwindow_as_diary)
     }
   }
 }