+++ R/iso8601chartime2POSIX.R
@@ -1,3 +1,3 @@
 iso8601chartime2POSIX = function(x, tz){
-  return(as.POSIXlt(x = x, format = "%Y-%m-%dT%H:%M:%S%z", tz = tz))
+  return(as.POSIXct(x = x, format = "%Y-%m-%dT%H:%M:%S%z", tz = tz))
 }